requirements.txt
aiogram~=2.14.3
psycopg2~=2.9.1
requests~=2.25.1
atlassian-python-api==3.12.1